Fix typing

+++ app/ml/training/train_analysis_model.py
@@ -11,6 +11,7 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
+
 import math
 import os
 import random
@@ -101,7 +102,7 @@ def fill_metric_stats(
 ) -> None:
     _, p_value = stats.f_oneway(baseline_model_metric_result, new_model_metric_results)
     if p_value is None or math.isnan(p_value):
-        p_value = 1.0
+        p_value = np.float64(1.0)
     info_dict["p_value"] = p_value
     mean_metric = np.mean(new_model_metric_results)
     baseline_mean_metric = np.mean(baseline_model_metric_result)
